FileName always removes pathnames.  This fixes the problem that temporary files didn't always get created in the outputdirectory but instead in some random directory (try calling Mx with a filename starting with ".." and shudder).
fmustopen now finds the ancestral directory that does exist and from
there creates subdirectories instead of starting from the beginning of
the path name until it finds a directory that does not exist.  This
fixes the problem that on Windows if you specify an absolute pathname
(i.e. one that starts with a "C:") fmustopen tried to mkdir("C:")
which fails.