TIMES 100: the future of query-intensive database kernels developed at CWI ==========================================================================
x100 is a database kernel that implements a standard relaional query algebra (scan,
select, project, aggr / todo: sort, join) using a volcano-like iterator
model. Its main divergent characteristics are:
- its execution primitives process value vectors iso one tuple at-a-time.
- it uses vertical decomposition throughout. (TODO:) Disk storage are compressed
  projection indices, maintained under updates using differntial lists.
- (TODO:) it tunes memory access: keeps random patterns inside the cpu-cache and boosts
  sequential access using a hardware abstraction layer for fast sequential transfers.

One could say that vectorized processing is similar to (multiplex) MIL processing.
While this is true, the crusial difference is that x100 is not penilized for
materializing result vectors, since vectors are variable and typically small, such
that all intermediate result vectors of an active query graph can be tuned to fit
the memory caches). As such, it does away with many existing Monet limitations in
terms of bandwidth wastage, among others with Monet's inability to efficiently execute
non-hardcoded multi-column expressions (such as compound selection conditions).

The challenge here in the x100 system is to decompose non-trivial query processing
algorithms in vectorized versions. Specifically, this means that program logic
inside if-then-else blocks, as well as while loops, that normally appear in query
processing algorithms (such as if-testing for hashing matches or while-looping over
a collision list) should be massaged out of the vectorized loop. all basic
steps should be taken on multiple tuples at-a-time, replacing if-testing by
vector size reduction (for this purpose there is basic support for indirect
value adressing: using an index vector to point to used slots in a data vector)
and replacing loops on tuples by loops on vectors.

Early experiments on Itanium2 machines have shown that vectorized execution on TPC-H
queries is 10 times faster than Monet (and 100 times faster than MySQL,BD2,Oracle)
an can even beat hand/hard-coded tuple-at-a-time iterator query plans. The main reason
for this is that branch preditions can be fully eliminated and the code has a very high
degree of independence, such that a wide-issue machine can get to its optimal IPC (>3!).

This code should be regarded as a preliminary prototype. In a future, we should
- rewrite: make the code standalone, convert to C++, split in various files, use m4 iso Mx.
- complete the functionality (varsized types, vector property propagation, sort, join)
- maybe even adapt the SQL frontend to also generate this algebra
- maybe even write a MIL executor on top of x100
- maybe even add transaction functionality