BIG CHANGE: Added support for 64 bit oid on 64-bit systems. This is done by using "typedef unsigned long oid;" which results on 32 bit oid on 32-bit systems and 64 bit oid on 64-bit machines. Many, many changes follow from this.
Also changes to make compilers happier.

Fix in module search/load code to only try to load a module if it has
actually been found.  This changes the expected error message for
non-existant modules.

All of this should have little effect on 32-bit machines.

To be done still:
define types for unsigned and signed integral types with the same size
as oid;
go through all remaining uses of "int" to see which type is really
meant.