Clean-up of the modules:
Some modules that are (currently) not considered "vital" for MonetDB are
moved from plain to contrib.

Two modules, "logger" and "perfex", are discontinued and hence removed
completely.

As opposed to those modules remaining in plain, the ones moved to contrib
are not "guaranteed" to be supported and kept up-to-date with all upcoming
changes in MonetDB.


For better traceability, I split this change in two step:

Step 1: Remove the respective files from plain and re-add their current
        version to contrib.

Step 2: Apply all the modifications necessary for this move.

Further, I tagged the cvs states before, between, and after these two steps
as "BeforePlain2Contrib", "WhilePlain2Contrib", and "AfterPlain2Contrib",
respectively.

This is Step 1.