Thanks to Walter's and Maurits' help, MonetDB 4.3 now also compiles and runs dynamically linked on AIX 5.1 both with GNU and IBM's native compiler; with the latter, even the 64-bits version does work.
There is only one (major) problem:
Some modules (mmath, vector, stats, radix, qt, calib) refuse to load.
"dlerror()" does not say anything; "errno" & "strerror()" say:
"#9: ..."
These are to be solved, once we are running regular tests on AIX
(unless someone volantiers to have a look ;-)

Details:
========
On AIX, shared libraries are called ".a", not ".so".
Basically, they are an archive with a shared object in it,
hence, dlopen needs to be given not only the name of the
archive, but also the name of the shared object in it.
Further, the dlopen mode needs to be or'ed with "RTLD_MEMBER".
Finally, the linker needs a "-brtl" switch.