fixed a DDImerge bug (cf. SF bug report #816672):
TMabort needs to undo changes not only in "dirty" BATs (BATdirty()),
but also in BATs with "diffs" (DELTAdirty()), i.e., saved/unloaded/swapped-out
BATs with differences that haven't been commit, yet.

Thanks to Niels for the tip where to look.

Actually, I guess, in the DELTAdirty()-case, we do not even have to load
the BAT and apply the BATundo, but could rather "simply" discard the delta
and revive the backup; however, I don't know (yet) how to do this,
hence we stick with the current solution for the time being --- it should
be correct though it might not be the most efficient way...

Also extended the repective test (tests/BugReports/Tests/no.089