If the set parameter is NULL, use an internal, static set variable instead.  Also, in that case read config options from the file pointed to by the MONET_CONFIG environment variable if it is set.
This change makes it possible to use this interface in
e.g. SQLConnect.  A Monet-aware ODBC application can, before calling
SQLConnect, call the various mo_* functions to set the options from
its arguments.  It then call SQLConnect which uses the options thus
set to find the SQL server.  The problem solved this way is that the
calling sequence of SQLConnect is fixed and does not allow passing any
extra pointers.