Reimplemented blocked streams.
We now no longer internally buffer everything until the call to
bs_flush (stream_flush).
The block boundaries are visible in the reader, so the caller has to
be able to deal with short reads.  Short reads don't indicate that the
other side did a flush at that point.
Calling bs_flush (stream_flush) causes a null block to be written
(i.e. just a count of 0).  This null block is returned to the caller
of bs_read (stream_read) as a 0 length buffer.  This can be used by
the caller to recognize when the sender flushed.
Also properly implemented bs_read_next (public function) to set the
*last parameter when it returns the last block before a flush.  The
calling sequence for bs_read_next was changed to include the size of
the buffer to be filled.

Finally, also added some comments to the implementation.