Added the use of memory mapped files backed by heap files on disk (even for new, temporary bats) to automatically tried possibilities for allocating BATs.
This means that on 64-bits platforms, you can now allocate BATs of obscene sizes.
Use at your peril though (if you try to process such beasts with a non-cacheable
access patterns: you *will* swap to death).

Also, on 32-bits platforms this sometimes allows you to further than your
configured swap space. That tends not to be a bottleneck anyways now, though
(the same 32-bits VM address space restrictions keep applying).



gdk.mx:
======

added GDKmallocmax(size,maxsize): a GDKmalloc that is able to pass a maximum size hint
to GDKvmalloc. This was done because we removed explicit GDKvmalloc use in Heaps.
The heap just uses GDKmalloc() or rather GDKvmallocmax() and if the block is large,
GDKmalloc itself decides to use anonymous virtual memory.

The decision to relegate this to GDKmalloc() was taken because GDKmalloc already did
do so anyway, and now we would like all hepas to have filenames available. In the
past STORE_MMAP without filename would indicate GDKvmalloc() memory. Now filenames
in heaps do not have this special meaning anymore.

gdk_bat.mx
==========

Add heap names in each new BAT descriptor. Thus, when a HEAPalloc or HEAPextend happens,
and the heap is gonna be really big, we know which file we could create to write into.

it is a kludge in the sense that it would have been better to add filenames to HEAPalloc()
HEAPextend() as parameters. However, these functions are also called indirectly, e.g. from
within extension atoms, so more APIs would need to be changed and more ADT implementations
fixed.

gdk_align.mx:
=============

add heapnames in VIEWreset, which makes a view independent of its parent. In BATmaterialize
which changes a void column into an oid column in the same non-view BAT we have a problem,
because those heap files are still in use. So we set the filenames to null, meaning that no
temporary memory mapped space is used.

this should be improved to increase scalability in BATmaterialize.

HEAPfree is now called always, as even if there is no heap, the filename should be freed.

The use of HASHnew also changes to put a filename in the new hash heap.


gdk_batop.mx
============

removed the use of HEAPsubcopy, which was disabled code anyway.

gdk_bbp.mx
==========


split BBPcacheit() into BBPcacheit() and BBPindex.

BBPindex finds a new slot in the BBP and sets it to used. Also puts the right name in the
provided BAT descriptor.

BBPcacheit is used to mark the fact that a BAT is loaded and available (in its preallocated
BBP slot).

This was needed because BATnewstorage needs to create filenames before entering the
BAT in the BBP (its last action , in the past). Ad a BAT filename depends on its BBP id.
This it can reserve the BBP id early with BBPinsert, even before the BAT is finished yet.

gdk_heap.mx
===========

HEAPsubcopy() was removed. Not used and terribly complicated to fix.

we now have the following possible 'kinds of heaps':
- STORE_MEM: allocated with GDKmallocmax(), extended withGDKrealloc() and freed with GDKfree().
             these now also use anonymous virtual memory under the hood if the block > GDK_mem_bigsize
- STORE_PRIV: 'safe' memory mapped files. Persistent BATs with writable heaps use these.
             Files are read with mmap(), but written with write().  They rename the heap file to a
             temporary .priv, such that writes can go to the reularly named file.
- STORE_MMAP: memory mapped files. For persistent BATs, these are used in read-only BATs.
	     However, writable non-persistent (temporary) BATs may now also use mmapped files.
             Corruption on a crash is not an issue, as non-persistent BATs do not survive crashes
             anyway (and their heap files are removed are removed in the cleanup scan at startup).

So, big temporary BATs use mmap, but the second deviation-from-normal is that they use .priv-named
filenames (just like STORE_PRIV). That is, if the BAT is changed into a persistent BAT and subsequently
saved, we write to an indepent file. After a crash, such files, when encountered, take precendece
over the .priv versions (so the newer, saved image is used).

Basically, the existing .priv infrastructre was reused for big temoporary mapped files.

gdk_search.mx
=============

Ugly, but before calling HASHnew, we should but a filename in the hash heap, such that when it
is huge (and by God those heaps can be) we can use the new memory mapped temporary heaps.

gdk_utils.mx
============

added GDKmallocmax(size,maxsize) that in case of size>GDK_mem_bigsize passes the maxsize
to GDKvmalloc.

GDKmalloc(size) = GDKmallocmax(size,size)

fixed GDKrealloc to properly try GDKvmrealloc() when applicable.

aggr.mx
=======

again, add the heap name before calling HASHnew().