Fixed a bug (hopefully) on windows, when the QUERYSIZE macro was defined, as on Windows. This macrol was illegally used also on vm memory blocks.
Turns out, doing that to get the block size is unavoidable. The solution has
been to abolish GDK_MEM_QUERYSIZE, and GDK_MEM_SIMPLE along the same line.

Now GDK memory blocks are *always* prefixed by a ssize_t that holds the
block size. Block sizes are always rounded up to 4-folds when requested.

If the stored prefix size is negative, we know the block is a vm block
(and prefixed by two ssize_t-s: 8 or 16 bytes, depending on the platform).

If the size is positive, it is a malloc block. How much extra bytes are prefixed?
- if the size is odd, it is prefixed by 4 bytes. This can only happen
  if GDK_MEM_MISALIGN is set. This indicates a libc where malloc can return
  blocks non-aligned to 8 bytes. This is *NOT* the default setting and should
  be set by configure. TODO TODO TODO!!
- it is prefixed by 8 bytes in all other cases.

Just before a block is freed, the 2-bit is set in its size field.
In this way, Monet can detect duplicate frees.

A final change, enabled by the separating out of GDK_MEM_MISALIGN, and
elimination of GDK_MEM_SIMPLE, is that in the default case we can use
again realloc (possibly saving a memcpy from src to dst). This was
previously impossible (and stays impossible in GDK_MEM_MISALIGN systems).