performance improvement for new heaps that are mmapped on temporary .priv files.
if such heaqps are made read-only, then they may be saved with an msync to X.priv
instead of a full save of X. this only can be done if X does not exist yet.

the code is not so clean (in other places), as a .priv extension in the filename
drives some behavior (the save method chosen in HEAPsave) rather than the
heap->storage field. Thus, the action we must take here is to remove the .priov
from the filename such that is treated as a normap STORE_MMAP heap.