BUGFIX [ 908219 ] [most probabaly] freeing non-allocated or already freed mem
this fixes the crash in tests/by_peter/Tests/GNATS-2736.milM

it may not fix the pther problem mentioned in the bug report.

cause was the tricky sharing of heaps in pump on the temporary
result bat. this was not undone fully before freeing this
temporary bat, leading to a free of the h->filename of the base BAT.