Implemented a new function in the MAPI library: mapi_next_result(Mapi mid).
Queries can result in multiple result set (i.e. result tables).  In
addition, the server may only return the first n rows for a result and
then start with the next result set (this is especially true for the
SQL server).

This is quite a big rewrite of parts of the MAPI library internals to
cope with these facts.  Another result of the rewrite is that the
purpose of internal functions has become much clearer.  Functions are
no longer used both to cache data which is being read ahead and to
return the next row to the application.  The read pointer in the cache
is now purely used to point at the next item to be returned to the
application.

Not a bug fix candidate.