Fix for bug 1011285.
The test whether we were at the end of file in doFileByLines was not
correct.  We need to check whether line==NULL and not whether
length==0, so we called mapi_query_done at the wrong time causing the
server to report an error on an empty line.

Also a slight efficiency change, don't send anything to the server for
an empty line.